Specify UID&GID
Due to the way file mapping from a volume into the container works, we need to set the correct file permissions on the files that will be used inside the container.

To be able to do this more reliably, we hard-code the GID and UID values, so that we can doccument which GID and UIDs will need to be given read, write, and execute permissions to these files.